# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title":"NSSMS API","version":"1.0.0","description":"Versioned API for the National School Sports Management System."},
  "servers":[{"url":"http://localhost:3000"}],
  "components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}}},
  "paths":{
...
    "/api/v1/admin/cards":{"get":{"security":[{"bearerAuth":[]}],"summary":"List issued enrollment cards","responses":{"200":{"description":"Card list"}}}},
    "/api/v1/admin/cards/issue":{"post":{"security":[{"bearerAuth":[]}],"summary":"Issue print-ready student, coach and representative cards after association acceptance","responses":{"201":{"description":"Issued"}}}},
...
  }
}